migrate support/tos to typescript
