+++ src/Settings/SettingsScreenNavigator.tsx
@@ -6,8 +6,6 @@ import {defineMessages, useIntl} from 'react-intl'
 import BiometricAuthScreen from '../../legacy/components/Send/BiometricAuthScreen'
 import BiometricsLinkScreen from '../../legacy/components/Settings/BiometricsLinkScreen'
 import LanguagePickerScreen from '../../legacy/components/Settings/ChangeLanguageScreen'
-import SupportScreen from '../../legacy/components/Settings/SupportScreen'
-import TermsOfServiceScreen from '../../legacy/components/Settings/TermsOfServiceScreen'
 import ToggleEasyConfirmationScreen from '../../legacy/components/Settings/ToggleEasyConfirmationScreen'
 import {defaultNavigationOptions, defaultStackNavigatorOptions} from '../../legacy/navigationOptions'
 import {SETTINGS_ROUTES, SETTINGS_TABS} from '../../legacy/RoutesList'
@@ -18,6 +16,8 @@ import {ChangePinScreen} from './ChangePin'
 import {ChangeWalletName} from './ChangeWalletName/ChangeWalletName'
 import {CustomPinScreen} from './CustomPin'
 import {RemoveWalletScreen} from './RemoveWallet'
+import {SupportScreen} from './Support'
+import {TermsOfServiceScreen} from './TermsOfService'
 import {WalletSettingsScreen} from './WalletSettings'
 
 /* eslint-disable @typescript-eslint/no-explicit-any */
