+++ src/Settings/Support/SupportScreen.stories.tsx
@@ -1,0 +1,8 @@
+import {storiesOf} from '@storybook/react-native'
+import React from 'react'
+
+import {SupportScreen} from './SupportScreen'
+
+storiesOf('SupportScreen', module).add('Default', () => {
+  return <SupportScreen />
+})
